+++ tests/multi_files/src/3085.php
@@ -1,0 +1,14 @@
+<?php
+
+function test_void_real() : void {
+}
+
+function expects_array_not_void(array $x) {
+    var_export($x);
+}
+
+function test_void_phpdoc() : void {
+    // Should not cause infinite loop with null_casts_as_any_type
+    echo strlen(test_void_real());
+    echo expects_array_not_void(test_void_real());
+}